requirements.txt
alpha_vantage
numpy
pandas
tensorflow
torch
transformers==4.37.2
scikit-learn
matplotlib
nltk
spacy
tqdm
PyYAML==5.3.1
jupyter
sentence-transformers
langchain
langchain_community
pypdf
openai
flask
autopep8
python-dateutil

